agrego boxes

+++ server/duck_action.cpp
@@ -7,7 +7,7 @@
 DuckAction::DuckAction(std::map<uint8_t, DuckPlayer>& map_personajes,
                        std::map<uint16_t,std::shared_ptr<Weapon>>& map_free_weapons,
                        std::map<uint16_t, std::unique_ptr<Bullet>>& map_bullets,
-                       uint16_t& id_balas, uint16_t& id_weapons, std::map<uint8_t, Protection>& map_helmet, std::map<uint8_t, Protection>& map_armor):
+                       uint16_t& id_balas, uint16_t& id_weapons, std::map<uint16_t, Protection>& map_helmet, std::map<uint16_t, Protection>& map_armor):
         map_personajes(map_personajes),
         map_free_weapons(map_free_weapons),
         map_bullets(map_bullets),
